# Use threads to clear the TT faster.