@MichaelB7 2018-05-15
Hash got slower to load when we changed the Hash  size away from a power of two back in December I believe. Will have to check this patch out , but glad to see somebody make an attempt to improve this.  It appears your patch does. 👊