@snicolet 2018-05-15
Interesting.

Could you provide the exact procedure you used to time the patch effect? What are the exact steps we can use to double-check the 50s and the 4s claim?