@RodrigoTjader 2018-05-15
The procedure I used was `time printf 'setoption name Threads value 20\nsetoption name Hash value 131072\n' | ./stockfish`

As soon as it finishes clearing the hash the UCI thread notices the input has been closed and exits.