@RodrigoTjader 2018-05-15
By the way, the power of two restriction is still mentioned in a comment
here:
https://github.com/official-stockfish/Stockfish/blob/master/src/tt.cpp#L31

On Tue, May 15, 2018 at 10:56 AM, Michael Byrne <notifications@github.com>
wrote:

> Hash got slower to load when we changed the Hash size away from a power of
> two back in December I believe. Will have to check this patch out , but
> glad to see somebody make an attempt to improve this. It appears your patch
> does. 👊
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/official-stockfish/Stockfish/pull/1601#issuecomment-389175966>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAE_8_SuirjlZD580ozTQx-DPN-PtVY2ks5tyt6WgaJpZM4T_lDj>
> .
>