@Hanamuke 2018-05-15
Some guy on stackoverflow was mentionning that write-through instruction existed for zeroing large chunk of memory. I guess it's an instruction that write in memory by bypassing the cache.