@vondele 2018-05-15
@RodrigoTjader I had been thinking about this as well, not so much for the time needed to zero the TT, but rather because there might be an important NUMA effect if this is done properly. A typical OS has a first touch placement policy for memory pages, and in master this zeroing by the master thread means that all TT memory is on the numa domain of that thread. That is likely suboptimal, and ideally on a NUMA system TT is spread over the NUMA domains (one supposes). Your patch might achieve this, with the caveat that we don't know the thread binding of the newly created threads. If we're lucky, they are spread evenly over the NUMA domains, but there is no guarantee. If we're unlucky the created threads are on a NUMA domain different from the search threads, and we could have a slowdown. It would be ideal if we could do the zeroing of memory using the search threads, as they are properly bound (at least on windows). Any chance using the search threads, or equivalently bound threads, is possible ? 

@MichaelB7 the time for clearing the hash got explicit once we moved to memset instead of calloc. The latter just moved that overhead into search. This has nothing todo with the power of two size of the hash.