@RodrigoTjader 2018-05-15
My first idea was using the search threads to do it, but that would be way more invasive, requiring more code in idle_loop with some verification to check if the thread was woken up for search or for clearing the hash, and also another barrier mechanism in order to block until all threads finish their work.

As that code would be much more complicated and confusing, moving some of the TT clearing logic into the Thread class, I went with this approach that felt much simpler. I have never messed with NUMA stuff, is there some way when creating a new thread to specify that it should be bound to the same domain as an existing thread? If so we could do that for the newly created threads, binding each one to the same domain as one of the existing search threads.

Even without this handling of NUMA the threads would likely be evenly spread, and if they aren't then the worse case is the same as it is without threads.