@RodrigoTjader 2018-05-15
Is https://github.com/official-stockfish/Stockfish/blob/master/src/thread.cpp#L100-L101 all that is done to bind threads to NUMA domains? Is just calling that with corresponding indexes for the clearing threads enough to get the same bindings?

Also, is the NUMA implementation only for Windows? From what I understand on other platforms 
bindThisThread() is just an empty function.