@vondele 2018-05-15
our messages crossed, but yes, I think that would be enough. And yes, there is only thread binding for windows for the time being.