@noobpwnftw 2018-05-15
Under extreme cases I would even suggest using node-local TT memory would be beneficial, it turns out that remote memory access not only come with latency but also bus congestion.

For this particular case I think it is fine not to bind threads to nodes but to use the number of NUMA nodes many of threads to perform the operation, the OS is smart enough to migrate things around automatically.