@RodrigoTjader 2018-05-15
@vondele yeah, that's a problem, I don't think most interfaces allow one to choose the order it sets the options. Maybe we could also switch the order in which stockfish prints them after the `uci` command in case some interface sets the options in the same order that the engine prints them. Always clearing hash when the thread count changes is also an option, but it would make the TT be cleared twice every time the engine is launched.

I have pushed a commit that includes a call to bindThisThread for further comments.