@noobpwnftw 2018-05-15
I have edited my previous comment on the Windows case, I think bindThisThread is favoring thread placement on the same node/group, but what we actually need is to solve the remote memory access case where the memory is remote to the thread that's zeroing it.