@vondele 2018-05-15
@noobpwnftw what this patch could address is the case where memory is all located in numa domain 0, but the search threads are on numa domain 0..X. This patch will spread the memory evenly over the domains, which ought to result in reduced congestion.. most importantly during search. Note that the patch only deals with the zeroing after allocation of the memory, so the pages will not be remote to a thread for the zeroing (assuming first touch policy).