@noobpwnftw 2018-05-15
Under Linux you can allocate memory on particular NUMA nodes and/or use "numactl --interleave=all" to control default memory placement policy, the default would be the node of the first thread that touches it.
How on Windows should we handle such case is unknown, but I would suspect the same unless explicitly allocated via VirtualAllocExNuma https://msdn.microsoft.com/en-us/library/windows/desktop/aa366891(v=vs.85).aspx.
Then, depending on where the memory pages were located, use the threads local to them?