@snicolet 2018-05-15
@RodrigoTjader 

Please try to keep everything as simple as possible, it is not because somebody says "it could also be beneficial to add NUMA binding" that you HAVE to add NUMA binding. We always must prove that there there is a measurable benefit before adding a feature, and even when there is a proven speed-up we must balance the added complexity to the speed-up.