@noobpwnftw 2018-05-15
Then I think the most reasonable and straightforward way of doing this is to allocate such memory with VirtualAllocExNuma and numa_tonode_memory(Linux equivalent). Otherwise the OS may still try to be smart and move things around.