@vondele 2018-05-16
I also repeated the zero-ing test with this patch against master, and can confirm a ~6.5x speedup, reducing from 5.5s to 0.85s the time for an ucinewgame command (on a dual socket 36 core system, using 28Gb hash). To test I run the following sequence of commands:

```
setoption name Threads value 36
setoption name Hash value  28000
ucinewgame
ucinewgame
ucinewgame
ucinewgame
ucinewgame
ucinewgame
ucinewgame
ucinewgame
ucinewgame
ucinewgame
ucinewgame
ucinewgame
ucinewgame
ucinewgame
ucinewgame
ucinewgame
ucinewgame
ucinewgame
ucinewgame
ucinewgame
quit
```
An measure the total runtime via:
```
cat commands.txt - | time ./stockfish
``` 
Time went from 110s to 17s